flie name to index.tsx,header tag are removed
