+++ src/App.css
@@ -33,6 +33,7 @@
   -webkit-background-clip: text;
   -webkit-text-fill-color: transparent;
 }
+/*Card-section */
 
 /* Header-slider */
