+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import Home from './Components/Home';
 import { Route, Routes } from 'react-router';
-import Navbar from './Components/Navbar/Navbar';
+import Navbar from './Components/Navbar';
 import { ROUTES } from './Constants/routeConstants';
 import About from './Components/About'
 import Footer from './Components/Footer';
