+++ src/Components/Navbar/index.tsx
@@ -10,13 +10,13 @@ const Navbar = () => {
                 <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
                     <span className="navbar-toggler-icon"></span>
                 </button>
-                <div className="collapse navbar-collapse" id="navbarNavDropdown">
+                <div className="collapse navbar-collapse" id="navbarNavDropdown bdNavbar">
                     <ul className="navbar-nav">
                         <li className="nav-item">
-                            <a className="nav-link " aria-current="page" href={ROUTES.HOME} >Home</a>
+                            <a className="nav-link" aria-current="page" href={ROUTES.HOME} >Home</a>
                         </li>
                         <li className="nav-item">
-                            <a className="nav-link" href={ROUTES.ABOUT}>About</a>
+                            <a className="nav-link" aria-current='true' href={ROUTES.ABOUT}>About</a>
                         </li>
                     </ul>
                 </div>
